refactor: promote CypherResultConverter to Application\Query\QueryResultNormalizer
Move the converter up to the Application layer so all query surfaces (Lua,
parser function, future REST handler) can share it. Behavior is preserved;
only namespace, class name, and file location change.

Add a refining @param annotation to convertAssociative() to satisfy PHPStan
at the new path.
